requirements.txt
# /mercado_livre_scraper/requirements.txt

Flask
requests
beautifulsoup4
supabase
pytz
Pillow
python-dotenv